wrap: Tests refactor.
