+++ cmd/wrap/main_test.go
@@ -22,98 +22,40 @@ import (
 )
 
 func TestInput(t *testing.T) {
-	in, err := ioutil.TempFile("", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer in.Close()
-
-	content := "trololo\n lololo\n lololo\n\n"
-	expected := "  trololo\n   lololo\n   lololo\n  \n"
-
-	_, err = io.WriteString(in, content)
-	if err != nil {
-		t.Fatal(err)
+	var inputTests = []struct {
+		content  string
+		indent   int
+		expected string
+	}{
+		{"trololo\n lololo\n lololo\n\n", 1, " trololo\n  lololo\n  lololo\n \n"},
+		{"trololo\n lololo\n lololo\n\n", 2, "  trololo\n   lololo\n   lololo\n  \n"},
 	}
 
-	_, err = in.Seek(0, os.SEEK_SET)
-	if err != nil {
-		t.Fatal(err)
-	}
+	for tc, tt := range inputTests {
+		in, err := ioutil.TempFile("", "")
+		if err != nil {
+			t.Fatal(err)
+		}
+		defer in.Close()
 
-	result := getInput(in)
+		if err != nil {
+			t.Fatal(err)
+		}
 
-	if result != expected {
-		t.Errorf("\"%s\" is not equal to \"%s\"", result, expected)
-	}
+		_, err = io.WriteString(in, tt.content)
+		if err != nil {
+			t.Fatal(err)
+		}
 
-}
+		_, err = in.Seek(0, os.SEEK_SET)
+		if err != nil {
+			t.Fatal(err)
+		}
 
-func TestKind(t *testing.T) {
-	yaml := `apiVersion: batch/v1
-kind: Job
-metadata:
-  name: pi
-spec:
-  template:
-    metadata:
-      name: pi
-    spec:
-      containers:
-      - name: pi
-        image: perl
-        command: ["perl",  "-Mbignum=bpi", "-wle", "print bpi(2000)"]
-      restartPolicy: Never`
-	kind, err := getKind(yaml)
-	if err != nil {
-		t.Fatal(err)
-	}
+		result := getInput(in, tt.indent)
 
-	if kind != "job" {
-		t.Errorf("Extracted kind should be \"job\", is %s", kind)
-	}
-}
-
-func TestWrap(t *testing.T) {
-	yaml := `  apiVersion: batch/v1
-  kind: Job
-  metadata:
-    name: pi
-  spec:
-    template:
-      metadata:
-        name: pi
-      spec:
-        containers:
-        - name: pi
-          image: perl
-          command: ["perl",  "-Mbignum=bpi", "-wle", "print bpi(2000)"]
-        restartPolicy: Never`
-
-	wrapped, err := getWrappedYaml(yaml, "name")
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := `apiVersion: appcontroller.k8s2/v1alpha1
-kind: Definition
-metadata:
-  name: name
-job:
-  apiVersion: batch/v1
-  kind: Job
-  metadata:
-    name: pi
-  spec:
-    template:
-      metadata:
-        name: pi
-      spec:
-        containers:
-        - name: pi
-          image: perl
-          command: ["perl",  "-Mbignum=bpi", "-wle", "print bpi(2000)"]
-        restartPolicy: Never`
-	if wrapped != expected {
-		t.Errorf("Wrapped doesn't match expected output\nExpected:\n%s\nAactual:\n%s", expected, wrapped)
+		if result != tt.expected {
+			t.Errorf("\"%s\" is not equal to \"%s\" in test case %d", result, tt.expected, tc+1)
+		}
 	}
 }
